Move to AVA for testing (#121)
